create step route
