+++ client/scenes/document/DocumentContainer.js
@@ -4,7 +4,6 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import {
   withRouter,
-  route,
   Switch,
   Route,
   Link,
