+++ client/scenes/json-schema-form-playground/components/Instructions.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export default props => <div>{props.content || "hello "}</div>;
