+++ client/scenes/json-schema-form-playground/components/TokenInformationForm.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export default props => <div>Token Information Form</div>;
