+++ client/scenes/json-schema-form-playground/components/index.js
@@ -1,0 +1,4 @@
+export { default as FormWizard } from "./FormWizard";
+export { default as Instructions } from "./Instructions";
+export { default as JsonSchemaForm } from "./JsonSchemaForm";
+export { default as TokenInformationForm } from "./TokenInformationForm";
